extract material slots in data grid premium
