+++ packages/grid/x-data-grid-premium/src/components/index.ts
@@ -1,4 +1,4 @@
 export * from './GridExcelExportMenuItem';
-export * from './icons';
+export * from '../material/icons';
 export * from './GridColumnMenuAggregationItem';
 export { GridColumnMenuGroupingItem } from './GridPremiumColumnMenu';
