+++ packages/grid/x-data-grid-premium/src/material/index.ts
@@ -1,0 +1,16 @@
+import { GridPremiumColumnMenu } from '../components/GridPremiumColumnMenu';
+import type { GridPremiumIconSlotsComponent } from '../models';
+import { GridWorkspacesIcon, GridGroupWorkIcon, GridFunctionsIcon } from './icons';
+
+const iconsSlots: GridPremiumIconSlotsComponent = {
+  ColumnMenuUngroupIcon: GridWorkspacesIcon,
+  ColumnMenuGroupIcon: GridGroupWorkIcon,
+  ColumnMenuAggregationIcon: GridFunctionsIcon,
+};
+
+const materialSlots = {
+  ...iconsSlots,
+  ColumnMenu: GridPremiumColumnMenu,
+};
+
+export default materialSlots;
